trying to get svg working

+++ spck.engine/src/main/java/spck/engine/ui/UIObjectPosition.java
@@ -1,10 +1,10 @@
 package spck.engine.ui;
 
-import org.joml.Vector2i;
+import org.joml.Vector2f;
 import spck.engine.Align;
 
 public class UIObjectPosition {
-    private final Vector2i position = new Vector2i();
+    private final Vector2f position = new Vector2f();
     private Align align;
 
     private UIObjectPosition() {
@@ -17,11 +17,11 @@ public class UIObjectPosition {
         return position;
     }
 
-    public void set(int x, int y) {
+    public void set(float x, float y) {
         position.set(x, y);
     }
 
-    public Vector2i get() {
+    public Vector2f get() {
         return position;
     }
